MovingObject moves in all directions with set speed

+++ include/Engine.h
@@ -13,7 +13,7 @@
 #include <iostream>
 
 namespace Engine {
-    extern double deltaTime;
+    extern float deltaTime;
 
     void ImGui();
     int Init();
